+++ package/Interface/CommunityShaders/Icons/Microsoft Icons/clear-cache.png
